requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
xgboost>=2.1
matplotlib>=3.9
seaborn>=0.13
jupyterlab>=4.2
pyyaml>=6.0
scipy>=1.13
torch>=2.3; platform_system != "Darwin" or platform_machine != "arm64"
